feat: adjust vite config, tsconfig, themeAtom, comment scss import, fix EditCluster, add plugin for paths and svgs
