+++ package.json
@@ -3,6 +3,7 @@
   "license": "Apache-2.0",
   "name": "@kyma-project/busola",
   "version": "0.0.1-rc.8",
+  "type": "module",
   "scripts": {
     "postinstall": "cd backend && npm install",
     "start": "concurrently -c cyan,green npm:busola npm:backend",
@@ -131,7 +132,9 @@
     "sass-loader": "^13.3.2",
     "typescript": "^4.8.4",
     "uuid": "^9.0.0",
-    "vite": "^5.4.3"
+    "vite": "^5.4.3",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^5.0.1"
   },
   "lint-staged": {
     "src/i18n/*.yaml": "yaml-sort --lineWidth=-1 -i ",
