+++ src/styles/index.scss
@@ -1,4 +1,4 @@
-@import '~@sap-theming/theming-base-content/content/Base/baseLib/sap_fiori_3/css_variables.css';
+// @import '~@sap-theming/theming-base-content/content/Base/baseLib/sap_fiori_3/css_variables.css';
 @import '../shared/components/ResourceGraph/mixins';
 
 @import './wizard-dialog.scss';
